+++ test/unit/fuzz-tools.test.js
@@ -222,6 +222,7 @@ import * as hematologyV229 from '../../lib/hematology-v229.js';
 import * as inflamV230 from '../../lib/inflam-v230.js';
 import * as prognosticV231 from '../../lib/prognostic-v231.js';
 import * as coagscoreV232 from '../../lib/coagscore-v232.js';
+import * as estimatorsV233 from '../../lib/estimators-v233.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -388,6 +389,7 @@ const MODULES = {
   'inflam-v230.js': inflamV230,
   'prognostic-v231.js': prognosticV231,
   'coagscore-v232.js': coagscoreV232,
+  'estimators-v233.js': estimatorsV233,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];